+++ pennylane_lightning/src/Gates.hpp
@@ -323,6 +323,23 @@ namespace Pennylane {
         void applyKernel(const StateVector& state, const std::vector<size_t>& indices, const std::vector<size_t>& externalIndices);
     };
 
+    class Unitary : public AbstractGate {
+    private:
+        const std::vector<CplxType> matrix;
+        const std::vector<double> real_matrix;
+    public:
+        static const std::string label;
+        static Unitary create(const std::vector<double>& parameters);
+        Unitary(const int numQubits, const std::vector<double>& real_matrix);
+        inline const std::vector<CplxType>& asMatrix() {
+            return matrix;
+        }
+        inline const std::vector<double>& asRealMatrix() {
+            return real_matrix;
+        }
+        void applyKernel(const StateVector& state, const std::vector<size_t>& indices, const std::vector<size_t>& externalIndices);
+    };
+
     /**
      * Produces the requested gate, defined by a label and the list of parameters
      *